Added update employee salary

+++ src/main/java/cs127/springappbe/Entities/Employee.java
@@ -31,7 +31,7 @@ public class Employee {
 
     private String Last_Name;
 
-    private long Salary;
+    private Double Salary;
 
     private String Account_Password;
 }
